need to add a function for get game

+++ Trivia_Yaniv_Shusterman/GameData.cpp
@@ -1,6 +1,6 @@
 #include "GameData.h"
 
-GameData::GameData(const Question currentQuestion, const unsigned int correctAnswerCount, const unsigned int wrongAnswerCount, const unsigned int averageAnswerTime)
+GameData::GameData(const Question currentQuestion, const unsigned int correctAnswerCount, const unsigned int wrongAnswerCount, const double averageAnswerTime)
 {
 	this->averageAnswerTime = averageAnswerTime;
 	this->correctAnswerCount = correctAnswerCount;
